+++ mimesis/builtins/ru.py
@@ -19,7 +19,7 @@ class RussiaSpecProvider(BaseSpecProvider):
     def __init__(self, seed: Optional[Seed] = None) -> None:
         """Initialize attributes."""
         super().__init__(locale=Locale.RU, seed=seed)
-        self._pull(self._datafile)
+        self._load_datafile(self._datafile)
         self._current_year = str(datetime.now().year)
 
     class Meta: