+++ mimesis/builtins/uk.py
@@ -17,7 +17,7 @@ class UkraineSpecProvider(BaseSpecProvider):
     def __init__(self, seed: Optional[Seed] = None) -> None:
         """Initialize attributes."""
         super().__init__(locale=Locale.UK, seed=seed)
-        self._pull(self._datafile)
+        self._load_datafile(self._datafile)
 
     class Meta:
         """The name of the provider."""