+++ mimesis/providers/date.py
@@ -24,7 +24,7 @@ class Datetime(BaseDataProvider):
         """Initialize attributes."""
         super().__init__(*args, **kwargs)
         self._datafile = "datetime.json"
-        self._pull(self._datafile)
+        self._load_datafile(self._datafile)
 
     class Meta:
         """Class for metadata."""