+++ mimesis/providers/finance.py
@@ -28,7 +28,7 @@ class Finance(BaseDataProvider):
         """
         super().__init__(*args, **kwargs)
         self._datafile = "finance.json"
-        self._pull(self._datafile)
+        self._load_datafile(self._datafile)
 
     class Meta:
         """Class for metadata."""