+++ mimesis/providers/food.py
@@ -18,7 +18,7 @@ class Food(BaseDataProvider):
         """
         super().__init__(*args, **kwargs)
         self._datafile = "food.json"
-        self._pull(self._datafile)
+        self._load_datafile(self._datafile)
 
     class Meta:
         """Class for metadata."""