+++ mimesis/providers/person.py
@@ -33,7 +33,7 @@ class Person(BaseDataProvider):
         """
         super().__init__(*args, **kwargs)
         self._datafile = "person.json"
-        self._pull(self._datafile)
+        self._load_datafile(self._datafile)
         self._store = {
             "age": 0,
         }