+++ mimesis/providers/text.py
@@ -21,7 +21,7 @@ class Text(BaseDataProvider):
         """
         super().__init__(*args, **kwargs)
         self._datafile = "text.json"
-        self._pull(self._datafile)
+        self._load_datafile(self._datafile)
 
     class Meta:
         """Class for metadata."""