+++ tests/test_providers/test_base.py
@@ -66,14 +66,14 @@ class TestBase(object):
     )
     def test_pull(self, locale, city):
         data_provider = BaseDataProvider(locale)
-        data_provider._pull("address.json")
+        data_provider._load_datafile("address.json")
         assert city in data_provider._data["city"]
 
     @pytest.mark.parametrize("locale", list(Locale))
     def test_pull_raises(self, locale):
         data_provider = BaseDataProvider(locale=locale)
         with pytest.raises(FileNotFoundError):
-            data_provider._pull("something.json")
+            data_provider._load_datafile("something.json")
 
     def test_extract(self, base_data_provider):
         dictionary = {"names": {"female": "Ariel", "male": "John"}}